Split pytests into separate file per class
